refactor : 테스트 패키지 구현코드와 동일하게 변경
