+++ src/main/java/model/participant/Participants.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import model.Deck.Deck;
+import model.deck.Deck;
 
 public final class Participants {
     private final List<Participant> participants;
