+++ src/test/java/model/deck/DeckTest.java
@@ -1,14 +1,10 @@
-package model;
+package model.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import model.Deck.Card;
-import model.Deck.CardRank;
-import model.Deck.CardSuit;
-import model.Deck.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
