+++ src/test/java/model/participant/PlayerTest.java
@@ -1,11 +1,10 @@
-package model;
+package model.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import model.Deck.Card;
-import model.Deck.CardRank;
-import model.Deck.CardSuit;
-import model.participant.Player;
+import model.deck.Card;
+import model.deck.CardRank;
+import model.deck.CardSuit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
