+++ src/test/java/model/result/ParticipantWinningResultTest.java
@@ -1,17 +1,15 @@
-package model;
+package model.result;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.Map;
-import model.Deck.Card;
-import model.Deck.CardRank;
-import model.Deck.CardSuit;
+import model.deck.Card;
+import model.deck.CardRank;
+import model.deck.CardSuit;
 import model.participant.Dealer;
 import model.participant.Participants;
 import model.participant.Player;
-import model.result.GameResult;
-import model.result.ParticipantWinningResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
